+++ akshare/futures/futures_zh_sina.py
@@ -695,7 +695,7 @@ if __name__ == "__main__":
     print(futures_zh_realtime_df)
 
     futures_zh_minute_sina_df = futures_zh_minute_sina(
-        symbol="V2201", period="5"
+        symbol="TF2009", period="1"
     )
     print(futures_zh_minute_sina_df)