+++ docs/changelog.md
@@ -34,6 +34,10 @@
 
 ## 更新说明
 
+1.7.28 fix: fix stock_info_sh_name_code interface
+
+    1. 修复 stock_info_sh_name_code 接口
+
 1.7.27 fix: fix stock_qsjy_em interface
 
     1. 修复 stock_qsjy_em 接口
@@ -1219,6 +1223,8 @@
     3. 修改 air_quality_hist 接口的请求日期格式，从 '2022-01-01' 为 '20220101'
 
 ## 版本更新说明
+
+1.7.28 fix: fix stock_info_sh_name_code interface
 
 1.7.27 fix: fix stock_qsjy_em interface